fastapi_backend/requirements.txt
fastapi==0.115.11
uvicorn==0.34.0
pydantic==2.10.6
python-dotenv==1.0.1
sqlalchemy==2.0.38
alembic==1.15.1
python-jose==3.4.0
passlib==1.7.4
rembg==2.0.64
numba>=0.61.0
requests==2.32.3
aiofiles==24.1.0
python-multipart==0.0.20
openai==1.65.5
matplotlib==3.10.1
pillow==11.1.0
numpy==1.26.4
httpx==0.28.1
# NLP Libraries for Semantic Analysis
spacy>=3.8.4
transformers>=4.49.0
torch==2.2.1